chore: 홈 화면 라우팅 경로 '/' > '/home'으로 변경
